# Use programmatic dependent launch in CUB merge sort